+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- bump to [cedar@v1.0.0-beta.4](https://github.com/Esri/cedar/releases/tag/v1.0.0-beta.4) and include arcgis-rest-js libs in vendor.js
+
 ## v1.0.0-beta.3
 ### Added
 - bump to [cedar@v1.0.0-beta.3](https://github.com/Esri/cedar/releases/tag/v1.0.0-beta.3)